+++ grocker/helpers.py
@@ -57,13 +57,13 @@ def render_template(template_path, output_path, context):
 def default_image_name(docker_image_prefix, release):
     req = pkg_resources.Requirement.parse(release)
     assert str(req.specifier).startswith('=='), "Only fixed version can use default image name."
-    return "{prefix}/{project}{extra_requirements}:{project_version}-{grocker_version}".format(
-        prefix=docker_image_prefix,
+    img_name = "{project}{extra_requirements}:{project_version}-{grocker_version}".format(
         project=req.project_name,
         extra_requirements='-' + '-'.join(req.extras) if req.extras else '',
         project_version=str(req.specifier)[2:],
         grocker_version=__version__,
     )
+    return '/'.join((docker_image_prefix, img_name)) if docker_image_prefix else img_name
 
 
 @contextlib.contextmanager